+++ src/PeriodicSystems.jl
@@ -36,8 +36,9 @@ export PeriodicTimeSeriesMatrix, HarmonicArray, FourierFunctionMatrix, PeriodicF
 export isperiodic, isconstant, iscontinuous, islti, set_period
 export mb03vd!, mb03vy!, mb03bd!, mb03wd!
 export ps
-export psaverage, psc2d, psmrc2d
+export psaverage, psc2d, psmrc2d, psteval
 export ps2fls, hr2bt, hr2btupd, phasemat, ps2frls, DiagDerOp, ps2ls
+export pspole, pszero
 
 abstract type AbstractDynamicalSystem end
 abstract type AbstractLTISystem <: AbstractDynamicalSystem end
@@ -52,5 +53,6 @@ include("types/PeriodicStateSpace.jl")
 include("ps.jl")
 include("conversions.jl")
 include("pslifting.jl")
+include("psanalysis.jl")
 include("psutils.jl")
 end